refactor: split OAuth pool account actions
