+++ .agents/plugins/opencode-aidevops/tests/test-oauth-pool-display.mjs
@@ -6,8 +6,16 @@ import assert from "node:assert/strict";
 
 import {
   formatDuration, formatAgo, poolActionCheck,
+  poolActionRemove, poolActionResetCooldowns,
+  poolActionAssignPending, poolActionSetPriority,
 } from "../oauth-pool-display.mjs";
 import { poolActionCheck as healthCheckAction } from "../oauth-pool-health-check.mjs";
+import {
+  poolActionRemove as accountRemoveAction,
+  poolActionResetCooldowns as accountResetCooldownsAction,
+  poolActionAssignPending as accountAssignPendingAction,
+  poolActionSetPriority as accountSetPriorityAction,
+} from "../oauth-pool-account-actions.mjs";
 
 test("display formatting keeps minute and hour output", () => {
   assert.equal(formatDuration(59_999), "0m");
@@ -17,4 +25,11 @@ test("display formatting keeps minute and hour output", () => {
 
 test("display module preserves the health-check action export", () => {
   assert.equal(poolActionCheck, healthCheckAction);
+});
+
+test("display module preserves account action exports", () => {
+  assert.equal(poolActionRemove, accountRemoveAction);
+  assert.equal(poolActionResetCooldowns, accountResetCooldownsAction);
+  assert.equal(poolActionAssignPending, accountAssignPendingAction);
+  assert.equal(poolActionSetPriority, accountSetPriorityAction);
 });
